requirements.txt
langchain==0.2.14
langchain-core==0.2.34
langchain-openai==0.1.22
langchain-community==0.2.12
langgraph==0.2.3
pydantic>=2.0.0
faiss-cpu==1.8.0.post1
chromadb==0.5.5
sentence-transformers==3.0.1
rank_bm25==0.2.2
python-dotenv==1.0.1
wikipedia==1.4.0
duckduckgo-search==6.2.10
colorama==0.4.6
